requirements.txt
click==7.1.2
cycler==0.10.0
Flask==1.1.2
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scipy==1.6.2
six==1.15.0
Werkzeug==1.0.1
